All model ids are Long now

+++ Laboratory2/src/main/java/com/musicshop/model/Cart.java
@@ -3,14 +3,14 @@ package com.musicshop.model;
 import java.time.LocalDateTime;
 
 public class Cart extends BaseModel<Long> {
-    private int customerID;
+    private Long customerID;
     private LocalDateTime dateCreated;
 
-    public int getCustomerID() {
+    public Long getCustomerID() {
         return customerID;
     }
 
-    public void setCustomerID(int customerID) {
+    public void setCustomerID(Long customerID) {
         this.customerID = customerID;
     }
